initial commit
This is the most recent version of starfield I have.

It's rewritten to TypeScript and using WebWorkers for calculations,
which gives a slight performance improvement. (If you don't wrap the
message received handler in requestAnimationFrame it gets even faster
but will also use ~double the CPU)

This is not using WebGL yet!